removed factored matrix functionality from shared util
