+++ transformer_lens/utilities/matrix.py
@@ -11,14 +11,6 @@ from jaxtyping import Float
 from transformer_lens.FactoredMatrix import FactoredMatrix
 
 
-def get_corner(tensor, n=3):
-    # Prints the top left corner of the tensor
-    if isinstance(tensor, torch.Tensor):
-        return tensor[tuple(slice(n) for _ in range(tensor.ndim))]
-    elif isinstance(tensor, FactoredMatrix):
-        return tensor[tuple(slice(n) for _ in range(tensor.ndim))].AB
-
-
 def composition_scores(
     left: FactoredMatrix, right: FactoredMatrix, broadcast_dims=True
 ) -> Union[
